refactor: split commands into helper modules
