+++ src-tauri/src/commands.rs
@@ -11,8 +11,8 @@ use std::{
 
 use dirs;
 
+use crate::python_helpers::conda_python;
 use crate::task_queue::{Task, TaskCommand, TaskQueue};
-use base64::{engine::general_purpose, Engine as _};
 use chrono::{Local, Utc};
 use rand::{thread_rng, Rng};
 use reqwest;
@@ -306,142 +306,6 @@ pub async fn comfy_stop() -> Result<(), String> {
         let _ = child.kill().await;
     }
     Ok(())
-}
-
-/* ==============================
-Python paths
-============================== */
-
-#[derive(Default, Serialize, Deserialize, Debug)]
-pub struct AppConfig {
-    pub python_path: Option<String>,
-    pub comfy_path: Option<String>,
-    pub sfz_convert_on_start: Option<bool>,
-    pub sfz_out_dir: Option<String>,
-}
-
-fn config_path() -> PathBuf {
-    let mut dir = dirs::home_dir().unwrap_or_else(|| PathBuf::from("."));
-    dir.push(".blossom");
-    let _ = fs::create_dir_all(&dir);
-    dir.push("config.json");
-    dir
-}
-
-fn load_config() -> AppConfig {
-    let path = config_path();
-    if let Ok(data) = fs::read_to_string(path) {
-        serde_json::from_str(&data).unwrap_or_default()
-    } else {
-        AppConfig::default()
-    }
-}
-
-fn save_config(cfg: &AppConfig) -> Result<(), String> {
-    let path = config_path();
-    let data = serde_json::to_string_pretty(cfg).map_err(|e| e.to_string())?;
-    fs::write(path, data).map_err(|e| e.to_string())
-}
-
-// Expose reading config to other modules (e.g., main.rs)
-pub fn get_config() -> AppConfig {
-    load_config()
-}
-
-#[tauri::command]
-pub async fn load_paths() -> Result<AppConfig, String> {
-    let path = resolve_python_path();
-    let mut cfg = load_config();
-    cfg.python_path = Some(path.to_string_lossy().to_string());
-    Ok(cfg)
-}
-
-#[tauri::command]
-pub async fn save_paths(
-    python_path: Option<String>,
-    comfy_path: Option<String>,
-    sfz_out_dir: Option<String>,
-) -> Result<(), String> {
-    let mut cfg = load_config();
-    if python_path.is_some() {
-        cfg.python_path = python_path;
-    }
-    if comfy_path.is_some() {
-        cfg.comfy_path = comfy_path;
-    }
-    if sfz_out_dir.is_some() {
-        cfg.sfz_out_dir = sfz_out_dir;
-    }
-    save_config(&cfg)
-}
-
-#[tauri::command]
-pub async fn set_sfz_convert_on_start(value: bool) -> Result<(), String> {
-    let mut cfg = load_config();
-    cfg.sfz_convert_on_start = Some(value);
-    save_config(&cfg)
-}
-
-fn default_python() -> PathBuf {
-    if cfg!(windows) {
-        PathBuf::from("python.exe")
-    } else {
-        PathBuf::from("python3")
-    }
-}
-
-/// Detect a python interpreter without consulting the user config.
-fn detect_python_path() -> PathBuf {
-    if let Ok(env_p) = std::env::var("BLOSSOM_PYTHON_PATH").or_else(|_| std::env::var("PYTHON")) {
-        if !env_p.trim().is_empty() {
-            return PathBuf::from(env_p);
-        }
-    }
-
-    let candidates: &[&str] = if cfg!(windows) {
-        &["python.exe", "python3.exe", "py.exe"]
-    } else {
-        &["python3", "python"]
-    };
-
-    for cand in candidates {
-        if let Ok(found) = which(cand) {
-            return found;
-        }
-    }
-
-    default_python()
-}
-
-/// Resolve the python interpreter path.
-/// Priority: user setting -> env var -> PATH search -> default.
-fn resolve_python_path() -> PathBuf {
-    let mut cfg = load_config();
-
-    if let Some(p) = &cfg.python_path {
-        if !p.trim().is_empty() {
-            return PathBuf::from(p);
-        }
-    }
-
-    let path = detect_python_path();
-    cfg.python_path = Some(path.to_string_lossy().to_string());
-    let _ = save_config(&cfg);
-    path
-}
-
-#[tauri::command]
-pub async fn detect_python() -> Result<String, String> {
-    Ok(detect_python_path().to_string_lossy().to_string())
-}
-
-/// Absolute path to python. Falls back to system default if unset.
-pub fn conda_python() -> PathBuf {
-    resolve_python_path()
-}
-
-pub fn conda_python_string() -> String {
-    conda_python().to_string_lossy().to_string()
 }
 
 fn default_comfy_path<R: Runtime>(app: &AppHandle<R>) -> PathBuf {
@@ -1924,126 +1788,6 @@ fn extract_intent(content: &str) -> String {
         "music" => "music".to_string(),
         _ => "chat".to_string(),
     }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct ShortSpec {
-    pub id: String,
-    pub title: String,
-    pub script: String,
-    pub audio_path: Option<String>,
-    pub visual_path: Option<String>,
-    pub export_path: Option<String>,
-    pub status: String,
-    pub created_at: String,
-}
-
-fn shorts_path() -> PathBuf {
-    let mut dir = dirs::home_dir().unwrap_or_else(|| PathBuf::from("."));
-    dir.push(".blossom");
-    let _ = fs::create_dir_all(&dir);
-    dir.push("shorts.json");
-    dir
-}
-
-#[tauri::command]
-pub async fn load_shorts() -> Result<Vec<ShortSpec>, String> {
-    let path = shorts_path();
-    if let Ok(data) = fs::read_to_string(path) {
-        serde_json::from_str(&data).map_err(|e| e.to_string())
-    } else {
-        Ok(vec![])
-    }
-}
-
-#[tauri::command]
-pub async fn save_shorts(specs: Vec<ShortSpec>) -> Result<(), String> {
-    let path = shorts_path();
-    if let Some(parent) = path.parent() {
-        let _ = fs::create_dir_all(parent);
-    }
-    let data = serde_json::to_string_pretty(&specs).map_err(|e| e.to_string())?;
-    fs::write(path, data).map_err(|e| e.to_string())
-}
-
-#[tauri::command]
-pub async fn generate_short(queue: State<'_, TaskQueue>, spec: ShortSpec) -> Result<u64, String> {
-    let label = format!("generate_short {}", spec.id);
-    let cmd = TaskCommand::GenerateShort { spec };
-    Ok(queue.enqueue(label, cmd).await)
-}
-
-/* ==============================
-Retro TV
-============================== */
-
-fn retro_tv_dir() -> PathBuf {
-    let mut dir = dirs::home_dir().unwrap_or_else(|| PathBuf::from("."));
-    dir.push(".blossom");
-    let _ = fs::create_dir_all(&dir);
-    dir
-}
-
-#[tauri::command]
-pub async fn save_retro_tv_video(data: String, ext: String) -> Result<String, String> {
-    let dir = retro_tv_dir();
-    // remove existing retro tv files
-    if let Ok(entries) = fs::read_dir(&dir) {
-        for entry in entries.flatten() {
-            let path = entry.path();
-            if path
-                .file_name()
-                .and_then(|n| n.to_str())
-                .map_or(false, |n| n.starts_with("retro_tv."))
-            {
-                let _ = fs::remove_file(path);
-            }
-        }
-    }
-    let path = dir.join(format!("retro_tv.{ext}"));
-    let b64 = data.split(',').last().unwrap_or_default();
-    let bytes = general_purpose::STANDARD
-        .decode(b64)
-        .map_err(|e| e.to_string())?;
-    fs::write(&path, bytes).map_err(|e| e.to_string())?;
-    Ok(path.to_string_lossy().to_string())
-}
-
-/* ==============================
-Video tools
-============================== */
-
-#[tauri::command]
-pub async fn loop_video(
-    input: String,
-    output_dir: String,
-    output_name: String,
-    hours: u64,
-    minutes: u64,
-    seconds: u64,
-) -> Result<String, String> {
-    let duration = hours * 3600 + minutes * 60 + seconds;
-    if duration == 0 {
-        return Err("duration must be greater than zero".into());
-    }
-    let out_path = Path::new(&output_dir).join(format!("{}.mp4", output_name));
-    let status = PCommand::new("ffmpeg")
-        .arg("-y")
-        .arg("-stream_loop")
-        .arg("-1")
-        .arg("-i")
-        .arg(&input)
-        .arg("-c")
-        .arg("copy")
-        .arg("-t")
-        .arg(duration.to_string())
-        .arg(&out_path)
-        .status()
-        .map_err(|e| e.to_string())?;
-    if !status.success() {
-        return Err(format!("ffmpeg exited with status {}", status));
-    }
-    Ok(out_path.to_string_lossy().to_string())
 }
 
 /* ==============================
