+++ src-tauri/tests/higgs_tts_errors.rs
@@ -1,8 +1,9 @@
 use std::{env, fs};
 
-use blossom_lib::commands::{higgs_tts, save_paths};
-use tempfile::tempdir;
+use blossom_lib::commands::higgs_tts;
+use blossom_lib::python_helpers::save_paths;
 use tauri::{test::mock_app, Manager};
+use tempfile::tempdir;
 use which::which;
 
 #[tokio::test]
@@ -23,7 +24,9 @@ async fn higgs_tts_reports_missing_python_and_script() {
     let handle = app.app_handle().clone();
     let missing = temp_home.path().join("no_python_here");
     env::set_var("BLOSSOM_PYTHON_PATH", missing.to_string_lossy().to_string());
-    let err = higgs_tts(handle, "hi".into(), "spk".into()).await.unwrap_err();
+    let err = higgs_tts(handle, "hi".into(), "spk".into())
+        .await
+        .unwrap_err();
     assert!(err.contains("Python not found"));
 
     // case: script missing
@@ -35,6 +38,8 @@ async fn higgs_tts_reports_missing_python_and_script() {
     env::set_current_dir(&temp_dir).unwrap();
     let app = mock_app();
     let handle = app.app_handle().clone();
-    let err = higgs_tts(handle, "hi".into(), "spk".into()).await.unwrap_err();
+    let err = higgs_tts(handle, "hi".into(), "spk".into())
+        .await
+        .unwrap_err();
     assert!(err.contains("Script not found"));
 }
